web-app/backend/requirements.txt
# Web App Backend Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0

# LangChain & LLM
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
langgraph==0.0.20

# Data Processing
pandas==2.2.2
numpy==1.24.3
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# API & HTTP
httpx==0.25.2
requests==2.32.3

# CORS & Security
python-cors==1.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Development
pytest==7.4.3
black==23.11.0
isort==5.12.0
